+++ app/src/main/java/com/pluu/webtoon/di/LoggerModule.kt
@@ -1,36 +1,16 @@
 package com.pluu.webtoon.di
 
-import com.pluu.webtoon.data.network.INetworkUseCase
-import com.pluu.webtoon.data.network.NetworkUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ApplicationComponent
-import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import timber.log.Timber
 import javax.inject.Singleton
 
 @Module
 @InstallIn(ApplicationComponent::class)
-object NetworkModule {
-
-    @Singleton
-    @Provides
-    fun provideNetworkUseCase(
-        okHttpClient: OkHttpClient
-    ): INetworkUseCase = NetworkUseCase(okHttpClient)
-
-    @Singleton
-    @Provides
-    fun provideOkHttpClient(
-        loggingInterceptor: HttpLoggingInterceptor
-    ): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor(loggingInterceptor)
-            .build()
-    }
-
+class LoggerModule {
     @Provides
     @Singleton
     fun provideLoggingInterceptor(): HttpLoggingInterceptor =
@@ -39,6 +19,6 @@ object NetworkModule {
                 Timber.tag("OkHttp").d(message)
             }
         }).apply {
-            level = HttpLoggingInterceptor.Level.BODY
+            level = HttpLoggingInterceptor.Level.BASIC
         }
 }
